bugfix in preprocess script to get "hypernyms" (ROOTs) from depparses some refactoring, etc. 	new file:   scripts/get_longman_hypernyms.sh 	modified:   src/def_to_dep.py 	modified:   src/preprocess.py 	new file:   src/utils.py
