+++ src/utils.py
@@ -1,0 +1,13 @@
+import os
+
+def ensure_dir(path):
+    if not os.path.exists(path):
+        os.mkdir(path)
+
+def batches(l, n):
+    """ Yield successive n-sized chunks from l.
+    (source: http://stackoverflow.com/questions/312443/
+    how-do-you-split-a-list-into-evenly-sized-chunks-in-python
+    """
+    for i in xrange(0, len(l), n):
+        yield l[i:i+n]
